修复多数据源问题；增强mongodb兼容性；还有上个版本改错了不好意思。

+++ src/nc/liat6/frame/db/connection/impl/ProxoolConnVarProvider.java
@@ -52,7 +52,7 @@ public class ProxoolConnVarProvider extends SuperConnVarProvider{
       Reflector.execute(ds,"setMaximumConnectionCount",new Class[]{int.class},new Object[]{setting.getMaximumConnectionCount()});
     }
     if(-1!=setting.getMaximumConnectionLifeTime()){
-      Reflector.execute(ds,"setMaximumConnectionLifeTime",new Class[]{int.class},new Object[]{setting.getMaximumConnectionLifeTime()});
+      Reflector.execute(ds,"setMaximumConnectionLifeTime",new Class[]{long.class},new Object[]{setting.getMaximumConnectionLifeTime()});
     }
     if(-1!=setting.getMinimumConnectionCount()){
       Reflector.execute(ds,"setMinimumConnectionCount",new Class[]{int.class},new Object[]{setting.getMinimumConnectionCount()});
